UserRepositoryのモックを作成

+++ go/src/MyPIPE/domain/repository/UserRepository.go
@@ -6,7 +6,7 @@ import (
 
 type UserRepository interface {
 	GetAll() []model.User
-	FindById(id int) *model.User
+	FindById(id uint64) *model.User
 	FindByToken(token string) *model.User
 	FindByEmail(email string) *model.User
 	SetUser(*model.User)
